fstracecheck: don't check 32bit on x86_64 host
Previously fstracecheck defaulted to check formatting for 64-bit and
32-bit architectures when executed on 64-bit host machine if -m option
wasn't specified.

However, it is better to default to checking formatting for only 64-bit
architecture when host machine is 64-bit. Checking 32-bit requires host
machine to have 32-bit build support installed which is not a reasonable
requirement when builds default to build only 64-bit libraries.